refactor App.java to use command design pattern
